% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanRCMadjustedCreateHourlyObsFile.R
\name{CanRCM4AdjustedCreateHourlyObs}
\alias{CanRCM4AdjustedCreateHourlyObs}
\title{Extracts all CRHM obs variables for a specified location and duration and constructs an obs file of hourly values.}
\usage{
CanRCM4AdjustedCreateHourlyObs(startDate = "1979-01-01",
  endDate = "2100-12-01", longitude = 0, latitude = 0,
  sunTimeOffset = 2, locationName = "",
  fileStr = "_CanRCM4_hist+fut_1979_2100", inDir = "./", outDir = "./",
  timezone = "etc/GMT+7", write3hour = FALSE, quiet = FALSE)
}
\arguments{
\item{startDate}{Optional. Beginning date of data to be extracted. A string formatted as "yyyy-mm-dd". The default value of \option{1979-01-01} is the beginning of the data.}

\item{endDate}{Optional. Beginning date of data to be extracted. A string formatted as "yyyy-mm-dd". The default value of \option{2100-12-01} is the end of the data.}

\item{longitude}{Required. The longitude of the point being sought. Valid values appear to be between -90 and -142, but the input value is \emph{not} checked for validity, in case the model extent changes.}

\item{latitude}{Required. The latitude of the point being sought. Valid values appear to be between 45 and 75, but the input value is \emph{not} checked for validity, in case the model extent changes.}

\item{sunTimeOffset}{Optional. The offset (in hours) is added to the solar time to convert it to local time. The default value \code{2} shifts the daily peak to occur at 2pm.}

\item{locationName}{Optional. Name for the location. This value is used to construct the name of the output file.}

\item{fileStr}{Optional. The name for the NetCDF .nc files containing the data, exclusive of their path and variable name.}

\item{inDir}{Optional. The path to the NetCDF .nc files. Default is the current directory.}

\item{outDir}{Optional. The path output file(s). Default is the current directory.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{write3hour}{Optional. Should the three-hour values be written to a .obs file? Default is \code{FALSE}.}

\item{quiet}{Optional. If \code{TRUE} then comments will be written to the screen as the input data is processed. Note that this function may take a long time to execute, so the default is the \emph{opposite} value used by most functions.}
}
\value{
If successful, writes \pkg{CRHMr} obs file(s) and returns \code{TRUE}. If unsuccessful, returns \code{FALSE}.
}
\description{
This function extracts the time series of 3-hourly values from a set of NetCDF files of CanRCM4 data, which has been bias-corrected using the WFDEI 3-hour reanalysis values. The values are stored at a spatial resolution of 0.125 degees, so the closest point to the specified location will be used. Note that because the reanalysis data omit February 29, the values returned by this function will hbe interplolate from February 28 and March 1 for leap days, except for precipitation values, which are set to zero. The variables are converted to values, and variable names, appropriate for CRHM when they are extracted. The three-hourly values of temeperature, wind speed
}
\examples{
\dontrun{CanRCM4AdjustedCreateHourlyObs(startDate = "1980-01-01", 
endDate = "1980-12-31", longitude = -101.704683, latitude = 50.845585, 
locationName = "SmithCreek")
}
}
\seealso{
\code{\link{CanRCM4adjustedGetNearestTimeseries}} \code{\link{qa2ea}}
}
