% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5deaccum.R
\name{ERA5deaccum}
\alias{ERA5deaccum}
\title{Deaccumulates ERA5 cumulative precipitation time series}
\usage{
ERA5deaccum(ERAobs, colnum = 1, quiet = TRUE, logfile = "")
}
\arguments{
\item{ERAobs}{Required. A \pkg{CRHMr} obs data frame of ERA 5 data, created by \code{ERAgetnearestTimeSeries}.}

\item{colnum}{Optional. The column number containing the values to be deaccumulated, not including the datetime. Default is column 1.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns an obs data frame containing the deaccumulated value. If unsuccessful, returns the value \code{FALSE}.
}
\description{
This function is used to deaccumulate variables stored as 24-hour cumulative values by ERA to 1-hour values. Note that this function only works for a single location, i.e. \emph{NOT} areal values.
}
\examples{
\dontrun{
deaccum <- ERA5deaccum(ERAobs)}

}
\author{
Kevin Shook
}
