% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAdailyArealPrecip.R
\name{ERAdailyArealPrecip}
\alias{ERAdailyArealPrecip}
\title{Calculates daily ERA areal precipitation}
\usage{
ERAdailyArealPrecip(ERAarealPrecip)
}
\arguments{
\item{ERAarealPrecip}{Required. The three-hour ERA areal precipitation as returned by \code{ERAgetArealPrecip}.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns a list containing the following: 
\tabular{ll}{
\bold{Name}\tab \bold{meaning}\cr
\code{daily_precip}\tab 3d array (lon x lat x date) of precip (mm)\cr
\code{lonres}\tab longitude resolution (\eqn{^\circ}{degrees})\cr
\code{latres}\tab latitude resolution (\eqn{^\circ}{degrees})\cr
\code{minLon}\tab mininum longitude of data(\eqn{^\circ}{degrees}W)\cr
\code{maxLon}\tab maximum longitude of data(\eqn{^\circ}{degrees}W)\cr
\code{minLat}\tab minimum latitude of data(\eqn{^\circ}{degrees})\cr
\code{minLat}\tab maximum latitude of data(\eqn{^\circ}{degrees})\cr
\code{date}\tab the date of each layer\cr
}
}
\description{
Calculates the daily total ERA precipitation, for all locations.
}
\examples{
\dontrun{daily_precip <- ERAdailyArealPrecip(threehour_precip)
}
}
\author{
Kevin Shook
}

