% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAdeaccum.R
\name{ERAdeaccum}
\alias{ERAdeaccum}
\title{Deaccumulates ERA cumulative time series}
\usage{
ERAdeaccum(ERAobs, colnum = 1, quiet = TRUE, logfile = "")
}
\arguments{
\item{ERAobs}{Required. A \pkg{CRHMr} obs dataframe of ERA data, created by \code{ERAgetnearestTimeSeries}.}

\item{colnum}{Optional. The column number containing the values to be deaccumulated, not including the datetime. Default is column 1.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used}
}
\value{
If successful, returns an obs dataframe containing the deaccumulated value. If unsuccessful, returns the value \code{FALSE}.
}
\description{
This function is used to deaccumulate variables stored as 12-hour cumulative values by ERA to 3-hour values. It is called by other functions, but may also be useful as a stand-alone functions. Note that this function only works for a single location, i.e. \emph{NOT} areal values.
}
\examples{
\dontrun{
deaccum <- ERAdeaccum(ERAobs)}

}
\author{
Kevin Shook
}
