% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAgetArealPrecip.R
\name{ERAgetArealPrecip}
\alias{ERAgetArealPrecip}
\title{Get ERA areal precipitation}
\usage{
ERAgetArealPrecip(ncdfFile, timezone = "", quiet = TRUE)
}
\arguments{
\item{ncdfFile}{Required. Name of the NetCDF file containing ERA data.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns a list containing the following: 
\tabular{ll}{
\bold{Name}\tab \bold{meaning}\cr
\code{precip}\tab 3d array (lon x lat x time) of precip (mm)\cr
\code{lonres}\tab longitude resolution (\eqn{^\circ}{degrees})\cr
\code{latres}\tab latitude resolution (\eqn{^\circ}{degrees})\cr
\code{minLon}\tab mininum longitude of data(\eqn{^\circ}{degrees}W)\cr
\code{maxLon}\tab maximum longitude of data(\eqn{^\circ}{degrees}W)\cr
\code{minLat}\tab minimum latitude of data(\eqn{^\circ}{degrees})\cr
\code{minLat}\tab maximum latitude of data(\eqn{^\circ}{degrees})\cr
\code{datetime}\tab the datetime of each time layer\cr
}
}
\description{
Extracts the areal precipitation for all time intervals from the ERA NetCDF file for a region. If the data are 3-hour values accumulated over 12-hour periods, then they will be deaccumulated, and negative precipitation will be set to zero
}
\examples{
\dontrun{threehour_precip <- 
ERAgetArealPrecip('_grib2netcdf-atls17-95e2cf679cd58ee9b4db4dd119a05a8d-szSVRK.nc', 
timezone='CST')
}
}
