% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAgetMultipleTimeseries.R
\name{ERAgetMultipleLocationTimeseries}
\alias{ERAgetMultipleLocationTimeseries}
\title{Extracts timeseries of ERA data nearest to multiple specified locations}
\usage{
ERAgetMultipleLocationTimeseries(
  ncdfFile,
  varName,
  siteNames,
  outDir = "",
  pointLons,
  pointLats,
  projection = "+proj=utm +zone=13 +ellps=WGS84",
  timezones = "",
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{ncdfFile}{Required. Name of the NetCDF file containing ERA data.}

\item{varName}{Required. Name of the NetCDF variable to extract.}

\item{siteNames}{Required. A vector containing the names of the sites. The names will be used for the output obs files.}

\item{outDir}{Optional. Directory to hold output files. If not specified, the current directory will be used.}

\item{pointLons}{Required. A vector containing decimal longitudes of desired locations. Note that the NetCDF longitude is 0-360\eqn{^\circ}{}, so add 360 to negative longitudes.}

\item{pointLats}{Required. A vector containing decimal latitudes of desired location.}

\item{projection}{Optional. Projection to be used to convert latitudes and longitudes to locations. Used for finding the nearest ERA gridpoint. The default, \option{+proj=utm +zone=13 +ellps=WGS84}, is \emph{only} valid for Western Canada. If you are processing data for the whole world, you can use the August Epicycloidal Projection which is \option{+proj=august +lon_0=90w} .}

\item{timezones}{Required. A vector containing the name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time. If there are fewer timezone values than variable names, the timezone will be recycled.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}, as it will list the actual latitude and longitude of the ERA tile.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns \code{TRUE} and all variables for all locations are written to \code{obs} files in the specified directory.
}
\description{
Reads a NetCDF file containing ERA reanalysis data and extracts the timeseries of the specified variable for each location. This is faster than calling the function \code{ERAgetNearestTimseries} for each location, as the reanalysis data are only read in once. Some of the the commonly-used variables are:
\tabular{llr}{
\bold{Parameter} \tab \bold{Units} \tab \bold{Variable}\cr
10 m eastward wind component \tab m/s \tab u10\cr
10 m northward wind component \tab m/s \tab v10\cr
2 metre temperature	\tab K \tab t2m \cr
2 metre dewpoint \tab K \tab d2m \cr
Downward surface solar radiation* \tab J/m\eqn{^2}{^2} \tab ssrd\cr
Downward surface thermal radiation* \tab J/m\eqn{^2}{^2} \tab strd\cr
Surface net solar radiation* \tab J/m\eqn{^2}{^2} \tab ssr \cr
Surface net thermal radiation* \tab J/m\eqn{^2}{^2} \tab str \cr
Total precipitation* \tab m of water \tab tp \cr
}
Parameters marked with an asterisk are \emph{cumulative} values, and must be deaccumulated using the deaccumERA function.
}
\examples{
\dontrun{
ERAgetMultipleLocationTimeseries('ssrd.nc', 'ssrd', s$`Station Names`,  
pointLons = s$Longitude, pointLats = s$Latitude, projection='+proj=august +lon_0=90w', 
timezones=s$timezone, quiet=FALSE)}
}
\seealso{
\code{\link{ERAgetNearestTimeseries}} \code{\link{ERAdeaccum}}
}
\author{
Kevin Shook
}
