% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAgetNearestTimeseries.R
\name{ERAgetNearestTimeseries}
\alias{ERAgetNearestTimeseries}
\title{Finds timeseries of ERA data nearest to specified location}
\usage{
ERAgetNearestTimeseries(
  ncdfFile,
  varName,
  pointLon,
  pointLat,
  projection = "+proj=utm +zone=13 +ellps=WGS84",
  timezone = "",
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{ncdfFile}{Required. Name of the NetCDF file containing ERA data.}

\item{varName}{Required. Name of the NetCDF variable to extract.}

\item{pointLon}{Required. Decimal longitude of desired location. Note that the NetCDF longitude is 0-360\eqn{^\circ}{}, so add 360 to negative longitudes.}

\item{pointLat}{Required. Decimal latitude of desired location.}

\item{projection}{Optional. Projection to be used to convert latitudes and longitudes to locations. Used for finding the nearest ERA gridpoint. The default, \option{+proj=utm +zone=13 +ellps=WGS84}, is \emph{only} valid for Western Canada. If you are processing data for the whole world, you can use the August Epicycloidal Projection which is \option{+proj=august +lon_0=90w} .}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}, as it will list the actual latitude and longitude of the ERA tile.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns a standard \pkg{CRHMr} dataframe containing the datetime and the extracted data, which are unpacked (i.e. the NetCDF multiplier and offset have been applied).
}
\description{
Reads a NetCDF file containing ERA reanalysis data and extracts the timeseries of the specified variable. Some of the the commonly-used variables are:
\tabular{llr}{
\bold{Parameter} \tab \bold{Units} \tab \bold{Variable}\cr
10 m eastward wind component \tab m/s \tab u10\cr
10 m northward wind component \tab m/s \tab v10\cr
2 metre temperature	\tab K \tab t2m \cr
2 metre dewpoint \tab K \tab d2m \cr
Downward surface solar radiation* \tab J/m\eqn{^2}{^2} \tab ssrd\cr
Downward surface thermal radiation* \tab J/m\eqn{^2}{^2} \tab strd\cr
Surface net solar radiation* \tab J/m\eqn{^2}{^2} \tab ssr \cr
Surface net thermal radiation* \tab J/m\eqn{^2}{^2} \tab str \cr
Total precipitation* \tab m of water \tab tp \cr
}
Parameters marked with an asterisk are \emph{cumulative} values, and must be deaccumulated using the deaccumERA function.
}
\examples{
\dontrun{
ssrd <- ERAgetNearestTimeseries('ssrd.nc', 'ssrd', 243.5, 51.69, timezone='Etc/GMT+7')
t2m <- ERAgetNearestTimeseries('t2m.nc', 't2m', 243.5, 51.69, timezone='Etc/GMT+7')
tp <- ERAgetNearestTimeseries('tp.nc', 'tp', 243.5, 51.69, timezone='Etc/GMT+7')
strd <- ERAgetNearestTimeseries('strd.nc', 'strd', 243.5, 51.69, timezone='Etc/GMT+7')
u10 <- ERAgetNearestTimeseries('u10v10.nc', 'u10', 243.5, 51.69, timezone='Etc/GMT+7')
v10 <- ERAgetNearestTimeseries('u10v10.nc', 'v10', 243.5, 51.69, timezone='Etc/GMT+7')
d2m <- ERAgetNearestTimeseries('d2m.nc', 'd2m', 243.5, 51.69, timezone='Etc/GMT+7')}
}
\seealso{
\code{\link{ERAdeaccum}}
}
\author{
Kevin Shook
}
