% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAhourlyAirtemp.R
\name{ERAhourlyAirtemp}
\alias{ERAhourlyAirtemp}
\title{Estimates hourly air temperatures from ERA 3 hourly values}
\usage{
ERAhourlyAirtemp(ERAt2m, t2mColnum = 1, method = "linear", quiet = TRUE,
  logfile = "")
}
\arguments{
\item{ERAt2m}{Required. The \pkg{CRHMr} obs dataframe of ERA t2m values. The values must not be deaccumulated, as the \code{deaccumERA} function is called by this function.}

\item{t2mColnum}{Optional. The column number containing the t2m values, not including the datetime. Default is column 1.}

\item{method}{Optional. The methods to be used for interpolation of the air temperature. Currently supported methods are \option{linear} and \option{spline}. The default is to use linear interpolation.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns an obs dataframe containing the interpolated hourly 2m air temperatures. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Interpolates ERA 3-hour instantaneous air temperatures to hourly values. The ERA air temperatures are first converted from K to \eqn{^\circ}{}C, if required.
}
\examples{
\dontrun{
hourlyTemps <- ERAhourlyAirtemp(t2m)}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{ERAgetNearestTimeseries}} \code{\link[CRHMr]{interpolate}}
}

