% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAhourlyLongwave.R
\name{ERAhourlyLongwave}
\alias{ERAhourlyLongwave}
\title{Estimates hourly incoming longwave radiation from ERA-Interim 3-hourly values or ERA-40 6-hourly values}
\usage{
ERAhourlyLongwave(ERAstrd, strdColnum = 1, ERAt2m, t2mColnum = 1,
  method = "linear", quiet = TRUE, logfile = "")
}
\arguments{
\item{ERAstrd}{Required. The \pkg{CRHMr} obs dataframe of ERA strd values. The values must not be deaccumulated, as the \code{ERAdeaccum} function is called by this function.}

\item{strdColnum}{Optional. The column number containing the strd values, not including the datetime. Default is column 1.}

\item{ERAt2m}{Required. The \pkg{CRHMr} obs dataframe of ERA t2m values. The values must not be deaccumulated, as the \code{deaccumERA} function is called by this function.}

\item{t2mColnum}{Optional. The column number containing the t2m values, not including the datetime. Default is column 1.}

\item{method}{Optional. The methods to be used for interpolation of the air temperature. Currently supported methods are \option{linear} and \option{spline}. The default is to use linear interpolation.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used}
}
\value{
If successful, returns an obs dataframe containing the interpolated hourly ERA incoming longwave radiation (Qli) in W/m\eqn{^2}{^2}, and the hourly 2m air temperatures (t) in \eqn{^\circ}{}C. If unsuccessful, returns the value \code{FALSE}.
}
\description{
This function is called \emph{after} using the function \code{getNearestERAtimeseries}. This function deaccumulates the ERA-Interim 12-hour cumulative values,using the function \code{ERAdeaccum} and interpolates the 3-hour values to hourly values, based on the extra-terrestrial hourly radiation. ERA-40 6-hourly instantaneous values are interpolated directly to hourly values.
}
\examples{
\dontrun{
hourlyQli <- ERAhourlyLongwave(ERAstrd=strd, ERAt2m=t2m)}
}
\references{
The value of the Stefan-Boltzmann constant was obtained from \url{http://physics.nist.gov/cgi-bin/cuu/Value?sigma}
}
\seealso{
\code{\link{ERAgetNearestTimeseries}}  \code{\link{ERAdeaccum}} \code{\link{ERAhourlyShortwave}}
}
\author{
Kevin Shook
}
