% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAhourlyPrecip.R
\name{ERAhourlyPrecip}
\alias{ERAhourlyPrecip}
\title{Etimates hourly precipitation from ERA cumulatve 3 hourly values}
\usage{
ERAhourlyPrecip(ERAtp, tpColnum = 1, quiet = TRUE, logfile = "")
}
\arguments{
\item{ERAtp}{Required. The \pkg{CRHMr} obs dataframe of ERA tp values. The values must not be deaccumulated, as the \code{deaccumERA} function is called by this function.}

\item{tpColnum}{Optional. The column number containing the tp values, not including the datetime. Default is column 1.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used}
}
\value{
If successful, returns an obs dataframe containing the interpolated hourly precipitation (p). If unsuccessful, returns the value \code{FALSE}.
}
\description{
This function is called \emph{after} using the function \code{getNearestERAtimeseries}. This function removes negative values, deaccumulates the 12-hour cumulative values to 3-hour values, using the function \code{deaccumERA}, and divides the results by 3 to give to hourly values.
}
\examples{
\dontrun{
hourlyP <- ERAhourlyPrecip(tp)}
}
\author{
Kevin Shook
}
\seealso{
\code{\link{ERAgetNearestTimeseries}}
}

