% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAhourlyVP.R
\name{ERAhourlyVP}
\alias{ERAhourlyVP}
\title{Calculates the hourly vapour pressure from the 3-hour ERA-Interim or 6 hour ERA-40 dew point temperature}
\usage{
ERAhourlyVP(ERAd2m, d2mColnum = 1, method = "linear", quiet = TRUE,
  logfile = "")
}
\arguments{
\item{ERAd2m}{Required. The \pkg{CRHMr} obs dataframe of ERA d2m values. The values must not be deaccumulated, as the \code{deaccumERA} function is called by this function.}

\item{d2mColnum}{Optional. The column number containing the d2m values, not including the datetime. Default is column 1.}

\item{method}{Optional. The methods to be used for interpolation of the dew point temperature. Currently supported methods are \option{linear} and \option{spline}. The default is to use linear interpolation.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns an obs dataframe containing the interpolated hourly 2m vapour pressures (ea) in kPA. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Interpolates ERA instantaneous dew point temperatures to hourly values. The ERA dew point temperatures are first converted from K to \eqn{^\circ}{}C, if required.
}
\examples{
\dontrun{
hourlyVP <- ERAhourlyVP(d2m)}
}
\seealso{
\code{\link{ERAgetNearestTimeseries}} \code{\link[CRHMr]{interpolate}} \code{\link{ERAhourlyAirtemp}}
}
\author{
Kevin Shook
}
