% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAhourlyWindspeed.R
\name{ERAhourlyWindspeed}
\alias{ERAhourlyWindspeed}
\title{Estimates hourly wind speeds from ERA 3 hourly wind vectors}
\usage{
ERAhourlyWindspeed(ERAu10, u10Colnum = 1, ERAv10, v10Colnum = 1,
  method = "linear", quiet = TRUE, logfile = "")
}
\arguments{
\item{ERAu10}{Required. The \pkg{CRHMr} obs dataframe of ERA u10 wind vector values. The values must not be deaccumulated, as the \code{deaccumERA} function is called by this function.}

\item{u10Colnum}{Optional. The column number containing the u10 values, not including the datetime. Default is column 1.}

\item{ERAv10}{Required. The \pkg{CRHMr} obs dataframe of ERA v10 wind vector values. The values must not be deaccumulated, as the \code{deaccumERA} function is vcalled by this function.}

\item{v10Colnum}{Optional. The column number containing the v10 values, not including the datetime. Default is column 1.}

\item{method}{Optional. The methods to be used for interpolation of the wind speeds. Currently supported methods are \option{linear} and \option{spline}. The default is to use linear interpolation.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns an obs dataframe containing the interpolated hourly 10m wind speeds (u10) in m/s and directions in degrees. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Estimates hourly wind speeds from ERA 3 hourly wind vectors
}
\examples{
\dontrun{
hourlyU <- ERAhourlyWindspeed(ERAu10=u10, ERAv10=v10)}
}
\seealso{
\code{\link{ERAgetNearestTimeseries}} \code{\link[CRHMr]{interpolate}}
}
\author{
Kevin Shook
}
