% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.R
\docType{data}
\name{land}
\alias{land}
\title{Land locations}
\format{
A dataframe with 67420 rows and 6 variables:
\describe{
  \item{Land}{number of the grid location}
  \item{Longitude}{longitude of the centre grid square}
  \item{Latitude}{latitude of the centre grid square}
  \item{Ht.m}{elevation of the centre grid squar}
  \item{glon}{x-location of the grid square}
  \item{glat}{y-location of the grid square}
}
}
\source{
This data was extracted from a file distributed with the WFD data, which can be downloaded from \url{http://www.eu-watch.org/}.
}
\usage{
land
}
\description{
A dataframe containing the information about the locations inside the WFD netCDF files. This data is used by the function \code{\link{WATCHcreateHourlyWFDobs}}.
}
\keyword{datasets}
