% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARRdownloadNetCDF.R
\name{NARRdownloadNetCDF}
\alias{NARRdownloadNetCDF}
\title{Downloads NARR NetCDF files}
\usage{
NARRdownloadNetCDF(interval = "daily", startYear = 1979, endYear = 1979,
  destination = ".", variable = "precip", quiet = FALSE)
}
\arguments{
\item{interval}{Optional. Time interval of NARR data. Can be '3h' or 'daily', which is the default.}

\item{startYear}{Optional. The first year to download. Default is 1979.}

\item{endYear}{Optional. The last year to download. Default is 1979.}

\item{destination}{Optional. The destination directory for the downloaded files. The default is the current directory.}

\item{variable}{Optional. The variable to be downloaded. Acceptable values are 'precip' (the default), 'temp', 'rh', 'wind', 'qli' and 'qli'.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. Because this function can be very slow to execute, the default value is \code{FALSE}, to provide information on the downloading.}
}
\value{
Writes the specified files to the destination directory. If successful, returns \code{TRUE}. If unsuccessful, returns {FALSE}.
}
\description{
Downloads NARR NetCDF files for a specified variable for a specified time step and range of years.
}
\examples{
\dontrun{
NARRdownloadNetCDF('3h', 1979, 2015)}
}

