% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa2ea.R
\name{qa2ea}
\alias{qa2ea}
\title{Converts specific humidity to vapour pressure.}
\usage{
qa2ea(qair, psurf)
}
\arguments{
\item{qair}{Required. Specific humidity (dimensionless).}

\item{psurf}{Required. Surface air pressure in Pa.}
}
\value{
If successful, returns the vapour pressure in kPa. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Converts specific humidity to vapour pressure.
}
\examples{
ea <- qa2ea(0.001, 101325)
}
\references{
R code for conversion of air pressure and absolute humidity was taken from project PEcAn The Predictive Ecosystem Analyzer \url{http://pecanproject.github.io}. The source code is available at \url{https://github.com/PecanProject/pecan/blob/master/modules/data.atmosphere/R/metutils.R.}
}
\author{
Kevin Shook
}
