% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reanalysis-package.R
\docType{package}
\name{Reanalysis-package}
\alias{Reanalysis-package}
\title{Contains functions to download and process reanalysis data}
\description{
The package contains functions to download and process reanalysis data. Functions are provided for ETA, NARR and WATCH datasets. Because there are so many functions and all of the datasets are slightly different, each function has the prefix of its dataset.
}
\references{
To cite \pkg{Reanalysis} in publications, use the command \code{citation('Reanalysis')} to get the current version of the citation.\cr
The CRHM program is described in:\cr
\cite{Pomeroy, John W, D M Gray, T Brown, N Hedstrom, W L Quinton, R J Granger, and S K Carey. 2007. \dQuote{The Cold Regions Hydrological Model : A Platform for Basing Process Representation and Model Structure on Physical Evidence}. Hydrological Processes 21 (19): 2650-2567.}\cr
The CRHM model may be downloaded from \url{http://www.usask.ca/hydrology/CRHM.php}.\cr
}
