% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WATCHcreateHourlyWFDEIobs.R
\name{WATCHcreateHourlyWFDEIobs}
\alias{WATCHcreateHourlyWFDEIobs}
\title{Creates a CRHM .obs file of hourly values from WATCH reanalysis data WFDEI files}
\usage{
WATCHcreateHourlyWFDEIobs(nc.location = "", startyear = 1979,
  endyear = 2012, lon = 0, lat = 0, precipType = "CRU",
  sunTimeOffset = 2, solarMethod = "simpleMaxSolar",
  interpolationMethod = "linear", obsFileName = "", timezone = "",
  quiet = TRUE, logfile = "")
}
\arguments{
\item{nc.location}{Required. A character string of the directory holding the WATCH WFDEI netCDf files. This is a file path WITHOUT a terminal slash, e.g. \option{z:\\WATCH\\WFDEI}}

\item{startyear}{Optional. Year to begin. Must be in the range \code{1979-2012}. Default is \code{1979}.}

\item{endyear}{Optional. Year to end. Must be in the range \code{1979-2012}. Default is \code{2012}.}

\item{lon}{Required. Decimal longitude to extract for.}

\item{lat}{Required. Decimal latitude to extract for.}

\item{precipType}{Optional. The precipitation type used; can be \option{CRU} (the default) or \option{GPCC}.}

\item{sunTimeOffset}{Optional. Number of hours that local noon is offset from solar noon. The default is 2 hours. This may be incorrect. It seems that WATCH data may not incorporate the temporal offset correctly - zero may work better.}

\item{solarMethod}{The method to be used for calculating the extra-terrestrial radiation. The default method is \option{simpleMaxSolar}. Note that this method is only valid for latitudes between 49 and 55\eqn{^\circ}{ }N. The other supported method is \option{PotSolarInst}, which requires the package \pkg{EcoHydRology} to be installed}

\item{interpolationMethod}{Optional. A vector containing the methods to be used for interpolation for each of the variables. Currently supported methods are \option{linear} and \option{spline}. The default is to use linear interpolation. If fewer methods than columns are specified, the methods are recycled.}

\item{obsFileName}{Required. Name of the .obs file to be created.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. You can use \option{Etc/GMT+6} or \option{Etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the value \code{TRUE} and writes the specified .obs file. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Extracts data from WATCH WDFEI netCDF files and builds a CRHM .obs file of hourly values containing \code{t}, \code{ea}, \code{u10}, \code{p} \code{Qsi} and \code{Qli}. All values are interpolated from 3 and 6 hur data. The windspeeds are at 10m, hence they are denoted as u10. Air temperatures are at 2m. The values for ea are computed from the atmospheric pressure (at 10m) and the absolute humidity (at 2m). Because the original NetCDF files are very large, this function typically runs very slowly. Also, because this function assembles and processes all of the data in memory, it can require a lot of RAM to execute.
}
\note{
Because this function can be slow, and uses a lot of memory, you may wish to run it repeatedly for short intervals. You can then use the function \code{appendObs} in \pkg{CRHMr} to join the files together. The time shifting results in obs files which do not begin and end on day boundaries, so you should use the function \code{trimObs} in \pkg{CRHMr} to trim the obs file - \emph{after} the final file has been assembled.
}
\examples{
\dontrun{
location <- 'z:\\data\\WATCH\\WFDEI'
obsName <- 'VermilionWATCH_WFDEI.obs'
WATCHcreateHourlyWFDEIobs(nc.location=location, 
startyear=1979, endyear=2001, lon=-111.9, lat=53.2, timezone='Etc/GMT+7', obsFileName=obsName)
# read in file and trim it
obs <- CRHMr::readObsFile(obsFile=obsName, timezone='Etc/GMT+7')
trimmedObs <- CRHMr::trimObs(obs)
CRHMr::writeObsFile(trimmedObs, obsFileName)}

}
\references{
\R code for conversion of air pressure and absolute humidity was taken from project PEcAn The Predictive Ecosystem Analyzer \url{http://pecanproject.github.io}. The source code is available at \url{https://github.com/PecanProject/pecan/blob/master/modules/data.atmosphere/R/metutils.R.}
}
\seealso{
\code{\link{WATCHcreateHourlyWFDobs}} \code{\link[CRHMr]{trimObs}} \code{\link[CRHMr]{appendObs}}
}
\author{
Kevin Shook.
}
