% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WATCHcreateHourlyWFDobs.R
\name{WATCHcreateHourlyWFDobs}
\alias{WATCHcreateHourlyWFDobs}
\title{Creates a CRHM .obs file of hourly values from WATCH reanalysis data WFD files}
\usage{
WATCHcreateHourlyWFDobs(nc.location = "", startyear = 1901,
  endyear = 2001, lon = 0, lat = 0, sunTimeOffset = 2,
  solarMethod = "simpleMaxSolar", interpolationMethod = "linear",
  obsFileName = "", timezone = "", quiet = TRUE, logfile = "")
}
\arguments{
\item{nc.location}{Required. A character string of the directory holding the WATCH WFD netCDf files. This is a file path WITHOUT a terminal slash, e.g. \option{z:\\WATCH\\WFD}.}

\item{startyear}{Optional. Year to begin. Must be in the range \code{1901-2001}. Default is \code{1901}.}

\item{endyear}{Optional. Year to end. Must be in the range \code{1901-2001}. Default is \code{2001}.}

\item{lon}{Required. Decimal longitude to extract for.}

\item{lat}{Required. Decimal latitude to extract for.}

\item{sunTimeOffset}{Optional. Number of hours that local noon is offset from solar noon. The default is 2 hours.}

\item{solarMethod}{The method to be used for calculating the extra-terrestrial radiation. The default method is \option{simpleMaxSolar}. Note that this method is only valid for latitudes between 49 and 55\eqn{^\circ}{ }N. The other supported method is \option{PotSolarInst}, which requires the package \pkg{EcoHydRology} to be installed}

\item{interpolationMethod}{Optional. A vector containing the methods to be used for interpolation for each of the variables. Currently supported methods are \option{linear} and \option{spline}. The default is to use linear interpolation. If fewer methods than columns are specified, the methods are recycled.}

\item{obsFileName}{Required. Name of the .obs file to be created.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. You can use \option{Etc/GMT+6} or \option{Etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the value \code{TRUE} and writes the specified .obs file. Each month's data is written as it is created. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Extracts data from WATCH WFD netCDF files and builds a CRHM .obs file of hourly values containing \code{t}, \code{ea}, \code{u10}, \code{p} \code{Qsi} and \code{Qli}. All values are interpolated from 3 and 6 hur data. The windspeeds are at 10m, hence they are denoted as u10. Air temperatures are at 2m. The values for ea are computed from the atmospheric pressure (at 10m) and the absolute humidity (at 2m). Because the original NetCDF files are very large, this function typically runs very slowly. Also, because this function assembles and processes all of the data in memory, it can require a lot of RAM to execute.
}
\examples{
\dontrun{
location <- 'z:\\data\\WATCH\\WFD'
obsFileName <- 'VermilionWATCH_WFD.obs'
WATCHcreateWFDobs(location, 1979, 2001, -111.9, 53.2, timezone='Etc/GMT+7', obsFileName=obsFileName)}
}
\references{
R code for conversion of air pressure and absolute humidity was taken from project PEcAn The Predictive Ecosystem Analyzer \url{http://pecanproject.github.io}. The source code is available at \url{https://github.com/PecanProject/pecan/blob/master/modules/data.atmosphere/R/metutils.R.}
}
\author{
Kevin Shook
}
