% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WATCHcreateWFDEIobs.R
\name{WATCHcreateWFDEIobs}
\alias{WATCHcreateWFDEIobs}
\title{Creates a CRHM .obs file of 3-hourly values from WATCH reanalysis data WFDEI files}
\usage{
WATCHcreateWFDEIobs(
  nc.location = "",
  startyear = 1979,
  endyear = 2012,
  lon = 0,
  lat = 0,
  houroffset = 0,
  obsFileName = "",
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{nc.location}{Required. A character string of the directory holding the WATCH WFDEI netCDf files. This is a file path WITHOUT a terminal slash, e.g. \option{z:\\WATCH\\WFDEI}}

\item{startyear}{Optional. Year to begin. Must be in the range \code{1979-2012}. Default is \code{1979}.}

\item{endyear}{Optional. Year to end. Must be in the range \code{1979-2012}. Default is \code{2012}.}

\item{lon}{Required. Decimal longitude to extract for.}

\item{lat}{Required. Decimal latitude to extract for.}

\item{houroffset}{Required. Number of hours that the local location is offset from UTC (GMT). Must be negative in the western hemisphere. For Mountain Standard Time, the offset is \code{-7} hours.}

\item{obsFileName}{Required. Name of the .obs file to be created.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the value \code{TRUE} and writes the specified .obs file. Each month's data is written as it is created. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Extracts data from WATCH WDFEI netCDF files and builds a CRHM .obs file of 3-hour data containing \code{t}, \code{ea}, \code{u10}, and \code{p} values. The output values can be interpolated to hourly using the function \code{HourlyWATCHobs}. The windspeeds are at 10m, hence they are denoted as u10. Air temperatures are at 2m. The values for ea are computed from the atmospheric pressure (at 10m) and the absolute humidity (at 2m).
}
\examples{
\dontrun{
location <- 'z:\data\WATCH\WFDEI'
obsFileName <- 'VermilionWATCH_WFDEI.obs'
WATCHcreateWFDEIobs(location, 1979, 2001, -111.9, 53.2, -7, obsFileName)}
}
\references{
R code for conversion of air pressure and absolute humidity was taken from project PEcAn The Predictive Ecosystem Analyzer \url{http://pecanproject.github.io}. The source code is available at \url{https://github.com/PecanProject/pecan/blob/master/modules/data.atmosphere/R/metutils.R.}
}
\author{
Kevin Shook.
}
