% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WATCHgroupWFDEIobs.R
\name{WATCHgroupWFDEIobs}
\alias{WATCHgroupWFDEIobs}
\title{Create obs files from WATCH WFDEI data for groups of sites}
\usage{
WATCHgroupWFDEIobs(
  siteFile,
  ncLocation,
  outputLocation,
  startyear = 1979,
  endyear = 2001,
  solarMethod = "PotSolarInst",
  interpolationMethod = "linear",
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{siteFile}{Required. A .csv file containing all of the variables required to describe the site locations. These are: 
\describe{
 \item{Name}{Name of site}
 \item{Land}{Land ID number}
 \item{Longitude}{Site longitude}
 \item{Latitude}{Site latitude}
 \item{glon}{Site glon value}
 \item{glat}{Site glat value}
 \item{timezone}{Time zone for site. Must be in Etc format, e.g. Etc/GMT+7}
 \item{SolarOffset}{Local offset of solar noon in hours} 
}}

\item{ncLocation}{Required. Location of all of the WATCH WFD files. Must have a trailing '\\' symbol.}

\item{outputLocation}{Required. Location for all of the output files. Must have a trailing '\\' symbol.}

\item{startyear}{Optional. Year to begin extraction. Default is 1901.}

\item{endyear}{Optional. Year to begin extraction. Default is 2001.}

\item{solarMethod}{The method to be used for calculating the extra-terrestrial radiation. The default method is \option{PotSolarInst}, which requires the package \pkg{EcoHydRology} to be installed. The other supported method is  \option{simpleMaxSolar}. Note that this method is only valid for latitudes between 49 and 55\eqn{^\circ}{ }N.}

\item{interpolationMethod}{Optional. A vector containing the methods to be used for interpolation for each of the variables. Currently supported methods are \option{linear} and \option{spline}. The default is to use linear interpolation. If fewer methods than columns are specified, the methods are recycled.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the value \code{TRUE} and writes the specified .obs file. Each month's data is written as it is created. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Extracts all variables from WFDEI NetCDF files and builds obs files for all specified locations
}
\examples{
\dontrun{
points <- './Reanalysis/testdata/WATCH_selected_points.csv'
outputLocation <- './Reanalysis/testdata/'
ncloc <- '//water.usask.ca/Centre/Reanalysis_data/WATCH/WFD'
WATCHgroupWFDEIobs(siteFile = points, ncLocation = ncloc, outputLocation=outputLocation, 
startyear = 1962, endyear = 1962))
}
}
\references{
R code for conversion of air pressure and absolute humidity was taken from project PEcAn The Predictive Ecosystem Analyzer \url{http://pecanproject.github.io}. The source code is available at \url{https://github.com/PecanProject/pecan/blob/master/modules/data.atmosphere/R/metutils.R.}
}
\author{
Kevin Shook
}
