% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WATCHhourlyObs.R
\name{WATCHhourlyObs}
\alias{WATCHhourlyObs}
\title{Interpolates 3 and 6 hour WATCH variables to hourly values}
\usage{
WATCHhourlyObs(infile = "", outfile = "", logfile = "")
}
\arguments{
\item{infile}{Required. Name of file created by CreateWFDobs or CreateWFDobs.}

\item{outfile}{Required. Hourly obs file to be created.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not use}
}
\value{
If successful, returns the value \code{TRUE} and writes the specified .obs file of hourly values. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Interpolates \code{t}, \code{u10}, \code{ea}, \code{qsi}, and \code{qli} values (6-hour in \code{WFD}, 3-hour in \code{WFDEI}) to hourly. The total \code{p} values are evenly divided over the hourly intervals. Outputs hourly \code{t}, \code{u10}, \code{ea}, \code{qsi}, \code{qli}, and \code{p}.
}
\examples{
\dontrun{
WATCHhourlyObs('VermilionWATCH_WFD.obs', 'VermilionWFDhourly.obs')
WATCHhourlyObs('VermilionWATCH_WFDEI.obs', 'VermilionWFDEIhourly.obs')}
}
\author{
Kevin Shook
}
