% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFDEI_GEM_CaPA_getNearestTimeseries.R
\name{WFDEI-GEM-WFDEI_GEM_CaPA_getNearestTimeseries}
\alias{WFDEI-GEM-WFDEI_GEM_CaPA_getNearestTimeseries}
\alias{WFDEI_GEM_CaPA_getNearestTimeseries}
\title{Finds timeseries of WFDEI-GEM-CaPA data nearest to specified location}
\usage{
WFDEI_GEM_CaPA_getNearestTimeseries(
  ncdfFile,
  varName,
  pointLon,
  pointLat,
  projection = "+proj=utm +zone=13 +ellps=WGS84",
  timezone = "",
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{ncdfFile}{Required. Name of the NetCDF file containing ERA data.}

\item{varName}{Required. Name of the NetCDF variable to extract.}

\item{pointLon}{Required. Decimal longitude of desired location.}

\item{pointLat}{Required. Decimal latitude of desired location.}

\item{projection}{Optional. Projection to be used to convert latitudes and longitudes to locations. Used for finding the nearest ERA gridpoint. The default, \option{+proj=utm +zone=13 +ellps=WGS84}, is \emph{only} valid for Western Canada. If you are processing data for the whole world, you can use the August Epicycloidal Projection which is \option{+proj=august +lon_0=90w} .}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}, as it will list the actual latitude and longitude of the ERA tile.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If unsuccessful, returns \code{FALSE}. If successful, returns a standard \pkg{CRHMr} dataframe containing the datetime and the extracted data, which are unpacked (i.e. the NetCDF multiplier and offset have been applied).
}
\description{
Reads a NetCDF file containing WFDEI-GEM-CaPA reanalysis data and extracts the timeseries of the specified variable. Some of the the commonly-used variables are:
\tabular{llr}{
\bold{Parameter} \tab \bold{Units} \tab \bold{Variable}\cr
40 m wind speed \tab m s-1 \tab wind_speed\cr
40 m air temperature	\tab K \tab ta \cr
40 m specific humidity \tab kg kg-1 \tab hus \cr
Surface precipitation\tab kg m-2 s-1 \tab pr \cr
Surface pressure\tab Pa \tab ps \cr
}
}
\seealso{
\code{\link{ERAdeaccum}}
}
\author{
Kevin Shook
}
