% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRF2obs.R
\name{WRF2Obs}
\alias{WRF2Obs}
\title{Converts a WRF netcdf file to obs}
\usage{
WRF2Obs(
  netCDFfile = "",
  obsfile = "",
  startDatetime = "2000-10-01 00:00",
  timezone = "Etc/GMT+7",
  trim = TRUE,
  quiet = TRUE,
  logfile = ""
)
}
\arguments{
\item{netCDFfile}{Required. The name of the netCDF file containing the WRF data.}

\item{obsfile}{Optional. If specified the values are written to the \code{obs} file.}

\item{startDatetime}{Optional. Date and time of first interval. Format is "yyyy-mm-dd hh:mm".}

\item{timezone}{Optional. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. You can use \option{Etc/GMT+6} or \option{Etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings.}

\item{trim}{Optional. If \code{TRUE} (the default) then the final output will be trimmed to CRHM day boundaries.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the value \code{TRUE} and (optionaly) writes the specified .obs file. If unsuccessful, returns the value \code{FALSE}.
}
\description{
Creates a \pkg{CRHMr} data frame of all values from a WRF netCDF file containing t, RH, p, and wind vectors for a \emph{single} location.
}
\examples{
\dontrun{w <- WRF2Obs("wrf2d_tquvr.nc")}
}
