% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WRFnearest2obs.R
\name{WRFnearest2obs}
\alias{WRFnearest2obs}
\title{Gets CRHM obs veriables closest to specified point}
\usage{
WRFnearest2obs(netCDFfile = "", obsfile = "", longitude = 0,
  latitude = 0, airPressure = 0, startDatetime = "2000-10-01 00:00",
  endDatetime = "2100-12-01 23:00", timezone = "", trim = TRUE,
  quiet = TRUE, logfile = "")
}
\arguments{
\item{netCDFfile}{Required. The name of the netCDF file containing the WRF data.}

\item{obsfile}{Optional. If specified the values are written to the \code{obs} file.}

\item{longitude}{Required. The longitude of the point being sought. The input value is \emph{not} checked for validity, in case the model extent changes.}

\item{latitude}{Required. The latitude of the point being sought. The input value is \emph{not} checked for validity, in case the model extent changes.}

\item{airPressure}{Optional. The mean air pressure in Pa. If set to zero (the default) then the air pressure within the NetCDf file will be used for converting the specific humidity to relative humidity. Because the air pressure may be incorrect or missing, you may wish to specify a mean air pressure to be used for the conversion. Note that if there is no air pressure within the NetCDF and you do not specify an air pressure value, this will trigger an error message, and the RH cannot be calculated.' @param endDatetime Optional. Ending date of data to be extracted. A string formatted as "yyyy-mm-dd hh:mm". The default value is \option{2100-12-01 23:00}.}

\item{startDatetime}{Optional. Beginning date of data to be extracted. A string formatted as "yyyy-mm-dd hh:mm". The default value is \option{2000-10-01 00:00}.}

\item{endDatetime}{Optional. Ending date of data to be extracted. A string formatted as "yyyy-mm-dd hh:mm". The default value is \option{2100-12-01 23:00}.}

\item{timezone}{Required. The name of the timezone of the data as a character string. This should be the timezone of your data, but omitting daylight savings time. Note that the timezone code is specific to your OS. To avoid problems, you should use a timezone without daylight savings time. Under Linux, you can use \option{CST} and \option{MST} for Central Standard or Mountain Standard time, respectively. Under Windows or OSX, you can use \option{etc/GMT+6} or \option{etc/GMT+7} for Central Standard and Mountain Standard time. DO NOT use \option{America/Regina} as the time zone, as it includes historical changes between standard and daylight savings time.}

\item{trim}{Optional. If \code{TRUE} (the default) then the final output will be trimmed to CRHM day boundaries.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used.}
}
\value{
If successful, returns the value \code{TRUE} and (optionaly) writes the specified .obs file. If unsuccessful, returns the value \code{FALSE}.

If successfl, returns either \code{TRUE} (if an obs file is specified) or a \pkg{CRHMr} obs data frame (if no obs file is specified). If unsucessful, retuns \code{FALSE}.
}
\description{
Creates a \pkg{CRHMr} data frame of all values from a WRF netCDF file for a specified location.
}
\examples{
\dontrun{ f <- "wrf2d_tquvr.nc"
r <- WRFnearest2obs(f, longitude=-115.27, 
latitude=52.03,  startDatetime = "1980-01-01 00:00", endDate="1980-12-31 23:00")
}
}
