% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERAhourlyShortwave.R
\name{ERAhourlyShortwave}
\alias{ERAhourlyShortwave}
\title{Estimates hourly incoming shortwave radiation from ERA 3 hourly values}
\usage{
ERAhourlyShortwave(ERAssrd, ssrdColnum = 1, latitude, sunTimeOffset = 2,
  solarMethod = "simpleMaxSolar", quiet = TRUE, logfile = "")
}
\arguments{
\item{ERAssrd}{Required. The \pkg{CRHMr} obs dataframe of ERA ssrd values. The values must not be deaccumulated, as the \code{deaccumERA} function is called by this function.}

\item{ssrdColnum}{Optional. The column number containing the ssrd values, not including the datetime. Default is column 1.}

\item{latitude}{Required. The latitude of the point at which the ssrd data is extracted. This value is used to calculate the extra-terrestrial incoming solar radiation.}

\item{sunTimeOffset}{Optional. Number of hours that local noon is offset from solar noon. The default is 2 hours.}

\item{solarMethod}{Optional. The method to be used for calculating the extra-terrestrial radiation. Default is \option{simpleMaxSolar}.}

\item{quiet}{Optional. Suppresses display of messages, except for errors. If you are calling this function in an \R script, you will usually leave \code{quiet=TRUE} (i.e. the default). If you are working interactively, you will probably want to set \code{quiet=FALSE}.}

\item{logfile}{Optional. Name of the file to be used for logging the action. Normally not used}
}
\value{
If successful, returns an obs dataframe containing the interpolated hourly ERA incoming shortwave radiation (Qsi). If unsuccessful, returns the value \code{FALSE}.
}
\description{
This function is called \emph{after} using the function \code{getNearestERAtimeseries}. This function removes negative values, deaccumulates the 12-hour cumulative values, using the function \code{ERAdeaccum} and interpolates the 3-hour values to hourly values, based on the extra-terrestrial hourly radiation.
}
\examples{
\dontrun{
hourlyQsi <- ERAhourlyShortwave(ssrd, latitude = 51.69)}

}
\seealso{
\code{\link{ERAgetNearestTimeseries}}  \code{\link{ERAhourlyLongwave}} \code{\link[CRHMr]{distributeQsi}}
}
\author{
Kevin Shook
}
